refactor(email): split email-sender into providers folder
Extract the single ~360-line email-sender.ts into a focused folder:
shared types (types.ts), helpers (shared.ts), one class per provider
under providers/, and the createEmailSender factory + re-export surface
in index.ts. The ./email-sender import path is unchanged (resolves to
index.ts), so no callers change. Pure code move, no behavior change.
