+++ worker/src/lib/email-sender/providers/noop.ts
@@ -1,0 +1,12 @@
+import type { EmailSender, SendEmailParams, SendEmailResult } from "../types";
+
+export class NoopSender implements EmailSender {
+  readonly provider = "none" as const;
+  async send(_: SendEmailParams): Promise<SendEmailResult> {
+    return { id: null, error: { message: "No email provider configured" } };
+  }
+
+  maxAttachmentBytes(): number {
+    return 0;
+  }
+}
