+++ worker/src/lib/email-sender/providers/resend.ts
@@ -1,0 +1,43 @@
+import { Resend } from "resend";
+import type { EmailSender, SendEmailParams, SendEmailResult } from "../types";
+import { toBase64 } from "../shared";
+
+export class ResendSender implements EmailSender {
+  readonly provider = "resend" as const;
+  private client: Resend;
+
+  constructor(apiKey: string) {
+    this.client = new Resend(apiKey);
+  }
+
+  async send(params: SendEmailParams): Promise<SendEmailResult> {
+    const result = await this.client.emails.send({
+      from: params.from,
+      to: params.to,
+      ...(params.cc && params.cc.length > 0 ? { cc: params.cc } : {}),
+      subject: params.subject,
+      html: params.html,
+      text: params.text,
+      headers: params.headers,
+      ...(params.attachments && params.attachments.length > 0
+        ? {
+            attachments: params.attachments.map((a) => ({
+              filename: a.filename,
+              content: toBase64(a.content),
+            })),
+          }
+        : {}),
+    });
+    if (result.error) {
+      return {
+        id: null,
+        error: { message: result.error.message ?? "Resend send failed" },
+      };
+    }
+    return { id: result.data?.id ?? null, error: null };
+  }
+
+  maxAttachmentBytes(): number {
+    return 25 * 1024 * 1024;
+  }
+}
