+++ worker/src/lib/email-sender/types.ts
@@ -1,0 +1,29 @@
+export interface SendEmailAttachment {
+  filename: string;
+  contentType: string;
+  /** Raw bytes. */
+  content: ArrayBuffer | Uint8Array;
+}
+
+export interface SendEmailParams {
+  from: string;
+  to: string;
+  /** Optional CC list — each entry can be a bare address or "Name <addr>". */
+  cc?: string[];
+  subject: string;
+  html: string;
+  text?: string;
+  headers?: Record<string, string>;
+  attachments?: SendEmailAttachment[];
+}
+
+export interface SendEmailResult {
+  id: string | null;
+  error: { message: string } | null;
+}
+
+export interface EmailSender {
+  provider: "resend" | "cloudflare" | "none" | "demo" | "bavimail";
+  send(params: SendEmailParams): Promise<SendEmailResult>;
+  maxAttachmentBytes(): number;
+}
